% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{cavtation_index}
\alias{cavtation_index}
\title{Cavitation Index (Regime)}
\usage{
cavtation_index(x, b, d, e, fls, sigma_value)
}
\arguments{
\item{x}{valve position}

\item{b}{steepness}

\item{d}{upper value}

\item{e}{the effective dose}

\item{fls}{liquid pressure recovery full open (max between fl and Flp/Fp)}

\item{sigma_value}{cavitation index (Reference downstream pressure P2)}
}
\value{
index 0 = free; 1 <- incipient; 2 <- constant; 3 <- maximum
}
\description{
Cavitation Index (Regime)
}
