% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpanetValve.R
\name{cylinder_param}
\alias{cylinder_param}
\title{calculation of the cylinder parameter}
\usage{
cylinder_param(valves, dn, d1, d2)
}
\arguments{
\item{valves}{tibble table}

\item{dn}{valve diameter (mm).}

\item{d1}{downstream pipe diameter (mm).}

\item{d2}{upstream pipe diameter (mm).}
}
\value{
calculation of the parameter as tibble table
}
\description{
calculation of the cylinder parameter
}
