% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{darcy_weisbach}
\alias{darcy_weisbach}
\title{Darcy–Weisbach equation}
\usage{
darcy_weisbach(flow, pipe_length, dn, roughness, temp = 15)
}
\arguments{
\item{flow}{cubic meter per second (m³/s)}

\item{pipe_length}{length of pipe  (m)}

\item{dn}{inner diameter of pipe  (m)}

\item{roughness}{internal roughness of the pipe in (m)}

\item{temp}{temperature in °C}
}
\value{
head loss (m)
}
\description{
In fluid dynamics, the Darcy–Weisbach equation is an empirical
equation, which relates the head loss, or pressure loss, due to friction
along a given length of pipe to the average velocity of the fluid flow
for an incompressible fluid.
The Darcy–Weisbach equation contains a dimensionless friction factor,
known as the Darcy friction factor. This is also variously called the
Darcy–Weisbach friction factor, friction factor, resistance coefficient,
or flow coefficient.
[https://en.wikipedia.org/wiki/Darcy%E2%80%93Weisbach_equation]
}
\examples{
darcy_weisbach( flow = 0.042,
                pipe_length = 970,
                dn = 0.150,
                roughness = 1.5e-6,
                temp = 14.5)

}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
