% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{dp_max}
\alias{dp_max}
\title{Differential pressure maximun between upstream and downstream pressure}
\usage{
dp_max(p1, fl, kv, dn, d1, d2, masl, temp)
}
\arguments{
\item{p1}{Gauge upstream pressure}

\item{fl}{liquid pressure recovery factor}

\item{kv}{Flow coefficient value in (m3/h).}

\item{dn}{valve diameter (m).}

\item{d1}{downstream pipe diameter (m).}

\item{d2}{upstream pipe diameter (m).}

\item{masl}{metres above sea level [m]}

\item{temp}{The temperature is in Celcius.}
}
\value{
DPmax (bar)
}
\description{
Differential pressure maximun between upstream and downstream pressure
}
