% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{dynamic_viscosity}
\alias{dynamic_viscosity}
\title{Water Dynamic Viscosity}
\usage{
dynamic_viscosity(temp = 15)
}
\arguments{
\item{temp}{is in °C}
}
\value{
Dynamic Viscosity (mPa*s)
}
\description{
Viscosity of the fluid is the internal resistance to flow. Fluid experiences
resistance over each layer of it when subjected to deformation.
viscosity is the measure of a fluid's resistance to flow
Dynamic vsicosity is the force required to maintain the fluid flow at a
particular rate. Dynamic viscosity is usually used when the fluid is
subjected to any external force. The unit of dynamic viscosit is Pa.s
where Pa is N/m^2 - SI unit of pressure
https://en.wikipedia.org/wiki/Temperature_dependence_of_viscosity
}
\examples{
dynamic_viscosity(14.5)

}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
