% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{ff}
\alias{ff}
\title{Ff Liquid critical pressure ratio factor}
\usage{
ff(temp = 15)
}
\arguments{
\item{temp}{is in °C}
}
\value{
Liquid critical pressure ratio factor Dimensionless
}
\description{
Ff is the liquid critical pressure ratio factor. This factor is
the ratio of the apparent vena contracta pressure at choked flow conditions
to the vapor pressure of the liquid at inlet temperature. [ISA-75.01.01-2007]
At vapor pressures near zero, this factor is 0.96.
}
\examples{
ff(15)
}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
