% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{fl_function}
\alias{fl_function}
\title{FL liquid pressure recovery factor}
\usage{
fl_function(fls, kv_kvs)
}
\arguments{
\item{fls}{liquid pressure recovery factor full open}

\item{kv_kvs}{relative flow coefficient as Array}
}
\value{
fl liquid pressure recovery factor for
}
\description{
for a control valve without attached fittings
}
