% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{flow_calculation}
\alias{flow_calculation}
\title{Flow Calculation}
\usage{
flow_calculation(pipe_length, dn, roughness, zeta, dp, temp = 15.6)
}
\arguments{
\item{pipe_length}{length of pipe  (m)}

\item{dn}{inner diameter of pipe  (m)}

\item{roughness}{internal roughness of the pipe in (m)}

\item{zeta}{loss coefficient}

\item{dp}{pressure diff.}

\item{temp}{temperature in °C}
}
\value{
flow in m3/s
}
\description{
Pressure Drop in Pipe with Losses (Determine flow)
}
