% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{flow_Kv}
\alias{flow_Kv}
\title{flow in function of the flow coefficient Kv}
\usage{
flow_Kv(p1, p2, kv, temp)
}
\arguments{
\item{p1}{Inlet Gauge pressure (bar).}

\item{p2}{Outlet Gauge pressure (bar).}

\item{kv}{Flow coefficient in (m³/h).}

\item{temp}{temperature is in Celcius.}
}
\value{
flow in (m³/h).
}
\description{
flow in function of the flow coefficient Kv
}
