% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{flow_through_valve}
\alias{flow_through_valve}
\title{Equations for incompressible fluids}
\usage{
flow_through_valve(kv, p1, p2, temp = 15, masl = 0, fl, fr = 1)
}
\arguments{
\item{kv}{Flow coefficient in m³/h}

\item{p1}{Inlet Absolute pressure}

\item{p2}{Outlet Absolute pressure}

\item{temp}{temperature in °C}

\item{masl}{metres above sea level [m]}

\item{fl}{Liquid pressure recovery factor of a control valve without attached fittings}

\item{fr}{Reynolds number factor.}
}
\value{
Volumetric flow rate in m³/h
}
\description{
The equations listed below identify the relationships between
flow rates, flow coefficients, related installation factors, and pertinent
service conditions for control valves handling incompressible fluids.
A sizing flow chart for incompressible fluids is given in Annex B. of the
ISA-75.01.01-2007 : Flow Equations for Sizing Control Valves.
}
\note{
"Absolute pressure" is zero-referenced against a perfect vacuum, using
an absolute scale, so it is equal to gauge pressure plus atmospheric pressure.
"Gauge pressure" is zero-referenced against ambient air pressure, so
it is equal to absolute pressure minus atmospheric pressure. [https://en.wikipedia.org/wiki/Pressure_measurement]
}
\examples{
 # Example 1: Incompressible flow – non-choked turbulent flow without attached fittings

 temp        <- 89                         # Water Inlet temperature °C
 density     <- water_density(temp)        # water density at 89°C degrees temperature in kg/m³
 pv          <- vapour_pressure(temp)*0.01 # Vapor pressure in bar
 k_viscosity <- kinematic_viscosity(temp)  # Kinematic viscosity in m²/s
 p1          <- 680*0.01                   # Upstream pressure in bar
 p2          <- 220*0.01                   # Dowstream pressure in bar
 dn          <- 0.150                      # Diamaeter in m
 patm        <- atm_pressure()             # Atm. Pressure
 ff_value    <- ff(temp)
 relat_density <- water_density(temp)/water_density(15)
 fl          <- 0.90                       # Liquid pressure recovery factor
 dp_max      <- (fl^2)*(p1-ff_value)

 flow_through_valve(kv = 165, p1 = 6.8, p2 = 2.2, temp = 89, fl = 0.9)

 # Example 2: Incompressible flow – choked flow without attached fittings
 fl <- 0.60   # Liquid pressure recovery factor
 dp_max      <- (fl^2)*(p1-ff_value)

 flow_through_valve(kv = 238, p1 = 6.8, p2 = 2.2, temp = 89, fl = 0.6)
}
