% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{flp}
\alias{flp}
\title{Flp Combined liquid pressure recovery factor}
\usage{
flp(kv, fl, dn, d1, d2)
}
\arguments{
\item{kv}{Flow coefficient in m³/h}

\item{fl}{Liquid pressure recovery factor of a control valve without attached fittings}

\item{dn}{diameter in meter [m]}

\item{d1}{Inlet diameter reducer only in meter [m]}

\item{d2}{Outlet diameter increaser only in meter [m]}
}
\value{
Product of the liquid pressure recovery factor of a valve with
attached fittings (no symbol has been identified) and the piping geometry
factor, dimensionless.
}
\description{
When a valve is installed with reducers or other attached
fittings, the liquid pressure recovery of the valve-fitting combination is
not the same as that for the valve alone. For calculations involving choked
flow, it is convenient to treat the piping geometry factor (fp) and the fl
factor for the valve-fitting combination as a single factor, flp.
}
\examples{
flp(kv = 7247.229, fl = 0.9, dn =0.5, d1 = 0.6, d2 = 0.6)

}
