% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{fr}
\alias{fr}
\title{Title}
\usage{
fr(reynolds_factor, Rev, problem_typ)
}
\arguments{
\item{reynolds_factor}{A \code{tibble} containing the values for the
calculation and plot of the Valve Reynolds Number Factor.}

\item{Rev}{Valve Reynolds number}

\item{problem_typ}{types of problems: "selection", "flow", "pressure"}
}
\value{
Valve Reynolds number factor
}
\description{
Title
}
