% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{friction_colebrook}
\alias{friction_colebrook}
\title{Colebrook–White equation (Darcy friction factor formula)}
\usage{
friction_colebrook(flow, dn, roughness, temp = 20)
}
\arguments{
\item{flow}{cubic meter per second (m³/s)}

\item{dn}{diameter in meter (m)}

\item{roughness}{roughness in (m)}

\item{temp}{temp is in °C}
}
\value{
the output from \code{\link{uniroot}} for the calculation of the
  root from the ecuation of Colebrook-White
}
\description{
The phenomenological Colebrook–White equation
(or Colebrook equation) expresses the Darcy friction factor f as a function
of Reynolds number Re and pipe relative roughness ε/Dh, fitting the data
of experimental studies of turbulent flow in smooth and rough pipes.
The equation can be used to (iteratively) solve for the Darcy–Weisbach
friction factor f.
[https://en.wikipedia.org/wiki/Darcy_friction_factor_formulae#Colebrook%E2%80%93White_equation]
}
\examples{
friction_colebrook(flow = 0.042, dn = 0.150, roughness = 1.5e-6, temp = 14.5)

}
