% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{friction_swamee}
\alias{friction_swamee}
\title{Friction loss in Pipe (Swanee-Jain)}
\usage{
friction_swamee(flow, dn, roughness, temp = 20)
}
\arguments{
\item{flow}{cubic meter per second (m³/s)}

\item{dn}{diameter in meter (m)}

\item{roughness}{internal roughness of the pipe in (m)}

\item{temp}{temperature is in °C}
}
\value{
Solve the ecuation of Colebrook-White with the ecuaton of
  Swanee-Jain (1976)
}
\description{
In fluid flow, friction loss (or skin friction) is the loss of
pressure or “head” that occurs in pipe or duct flow due to the effect of
the fluid's viscosity near the surface of the pipe or duct.
[https://en.wikipedia.org/wiki/Friction_loss]
}
\examples{
friction_swamee(flow = 0.042, dn = 0.150, roughness = 1.5e-6, temp = 14.5)

}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
