% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{kv}
\alias{kv}
\title{Valve flow coefficient}
\usage{
kv(p1, p2, flow, temperature = 15.6)
}
\arguments{
\item{p1}{Inlet Absolute pressure [bar]}

\item{p2}{Outlet Absolute pressure [bar]}

\item{flow}{flow in m³/h}

\item{temperature}{Inlet water temperature in °C}
}
\value{
kv Flow coefficient in m³/h
}
\description{
The valve flow coefficient Kv is defined as the number
of cubic meters per hour of water that will flow through
a control valve at a specified position of the control valve (travel) h
with a differential pressure Delta P (p1-p2) across it.
}
\examples{
kv(2, 1, 200)
}
