% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{kv_value}
\alias{kv_value}
\title{Kv Value}
\usage{
kv_value(dn, zeta)
}
\arguments{
\item{dn}{diameter in meter [m]}

\item{zeta}{dimensionless quantity}
}
\value{
kv value
}
\description{
Kv Value
}
\examples{

}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
