% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curv.R
\name{plot_fl}
\alias{plot_fl}
\title{Plot liquid pressure recovery factor}
\usage{
plot_fl(b, d, e, fls, cylindertyp)
}
\arguments{
\item{b}{steepness}

\item{d}{upper value}

\item{e}{the effective dose}

\item{fls}{liquid pressure recovery full open (max between fl and Flp/Fp)}

\item{cylindertyp}{Name of the control characteristic}
}
\value{
ggplot graphic. of the liquid pressure recovery factor
}
\description{
Plot liquid pressure recovery factor
}
