% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curv.R
\name{plot_kv}
\alias{plot_kv}
\title{Plot Flow coefficient Kv Value}
\usage{
plot_kv(b, d, e, dn, zvs, cylindertyp)
}
\arguments{
\item{b}{steepness}

\item{d}{upper value}

\item{e}{the effective dose}

\item{dn}{dn diameter in meter (m)}

\item{zvs}{Resistance Coefficient Zeta full open}

\item{cylindertyp}{Name of the control characteristic}
}
\value{
ggplot graphic. Flow coefficient Kv Value
}
\description{
Plot Flow coefficient Kv Value
}
