% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curv.R
\name{plot_kv_kvs}
\alias{plot_kv_kvs}
\title{Plott the Valve flow coefficient}
\usage{
plot_kv_kvs(b, d, e, cylindertyp)
}
\arguments{
\item{b}{steepness}

\item{d}{upper value}

\item{e}{the effective dose}

\item{cylindertyp}{Name of the control type}
}
\value{
ggplot graf.
}
\description{
Plott the Valve flow coefficient
}
