% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curv.R
\name{plot_sigma}
\alias{plot_sigma}
\title{Plot sigma for Incipient, Constant and Maximum cavitation}
\usage{
plot_sigma(b, d, e, fls, cylindertyp)
}
\arguments{
\item{b}{steepness}

\item{d}{upper value}

\item{e}{the effective dose}

\item{fls}{liquid pressure recovery full open (max between fl and Flp/Fp)}

\item{cylindertyp}{Name of the control characteristic}
}
\value{
ggplot graphic. of the sigma for Incipient, Constant and Maximum cavitation
}
\description{
Plot sigma for Incipient, Constant and Maximum cavitation
}
