% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curv.R
\name{plot_zv}
\alias{plot_zv}
\title{Plot Zeta Value Curve}
\usage{
plot_zv(b, d, e, zvs, cylindertyp)
}
\arguments{
\item{b}{steepness}

\item{d}{upper value}

\item{e}{the effective dose}

\item{zvs}{Resistance Coefficient Zeta full open}

\item{cylindertyp}{Name of the control characteristic}
}
\value{
ggplot graphic. Zeta Value Curve
}
\description{
Plot Zeta Value Curve
}
