% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_curv.R
\name{plotly_kv}
\alias{plotly_kv}
\title{Plottly Flow Coefficient (Kv) of the valve}
\usage{
plotly_kv(valves, i)
}
\arguments{
\item{valves}{tibble with the valves factors}

\item{i}{cylinder order}
}
\value{
plotly graph
}
\description{
Plottly Flow Coefficient (Kv) of the valve
}
