% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_curv.R
\name{plotly_kv_kvs}
\alias{plotly_kv_kvs}
\title{Plotly the Valve flow coefficient}
\usage{
plotly_kv_kvs(valves, i)
}
\arguments{
\item{valves}{tibble with the valves factors}

\item{i}{cylinder order}
}
\value{
plot in gglot format
}
\description{
Plotly the Valve flow coefficient
}
