% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_curv.R
\name{plotly_sigma}
\alias{plotly_sigma}
\title{Plottly Cavitation Coefficients}
\usage{
plotly_sigma(valves, i)
}
\arguments{
\item{valves}{tibble with the valves factors}

\item{i}{cylinder order}
}
\value{
plotly graph
}
\description{
Plottly Cavitation Coefficients
}
