% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_curv.R
\name{plotly_zv}
\alias{plotly_zv}
\title{Plotly Zeta Value Curve}
\usage{
plotly_zv(valves, i)
}
\arguments{
\item{valves}{tibble with the valves factors}

\item{i}{cylinder order}
}
\value{
plotly graph
}
\description{
Plotly Zeta Value Curve
}
