% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpanetValve.R
\name{points_cloud_param}
\alias{points_cloud_param}
\title{Calculation of the factors for the points cloud}
\usage{
points_cloud_param(base_data, dn, masl, temp)
}
\arguments{
\item{base_data}{tibble table}

\item{dn}{valve diameter (mm).}

\item{masl}{meters above sea level (m).}

\item{temp}{The water temperature is in Celsius.}
}
\value{
calculation of the parameter as tibble table
}
\description{
Calculation of the factors for the points cloud
}
