% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{resistance_coefficient}
\alias{resistance_coefficient}
\title{Resistance coefficients of all fittings attached to the control valve}
\usage{
resistance_coefficient(dn, d1, d2)
}
\arguments{
\item{dn}{valve diameter (m).}

\item{d1}{downstream pipe diameter (m).}

\item{d2}{upstream pipe diameter (m)}
}
\value{
Resistance coefficient.
}
\description{
The algebraic sum of all effective resistance coefficients of
all fittings attached to the control valve.
}
