% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{reynolds_number}
\alias{reynolds_number}
\title{Reynolds number}
\usage{
reynolds_number(flow, dn, temp = 20)
}
\arguments{
\item{flow}{cubic meter per second (m³/s)}

\item{dn}{diameter in meter (m)}

\item{temp}{temp is in °C}
}
\value{
reynolds number dimensionless quantity
}
\description{
The Reynolds number (Re) is an important dimensionless quantity in fluid
mechanics used to help predict flow patterns in different fluid flow situations.
}
\examples{


}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
