% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{Reynolds_valve}
\alias{Reynolds_valve}
\title{Valve Reynolds number}
\usage{
Reynolds_valve(flow, kv, dn, d1, d2, temp, fl, fd)
}
\arguments{
\item{flow}{flow in m³/h}

\item{kv}{Flow coefficient in m³/h}

\item{dn}{diameter in millimeter (mm).}

\item{d1}{Inlet diameter reducer only in millimeter (mm).}

\item{d2}{Outlet diameter increase only in millimeter (mm).}

\item{temp}{temperature is in Celsius.}

\item{fl}{Liquid pressure recovery factor of a control valve without attached fittings}

\item{fd}{Valve style modifier (1.0 fuer Ventile mit V-fourmigen Drosselquerschnitt)}
}
\value{
valve Reynolds number
}
\description{
Valve Reynolds number
}
