% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curv.R
\name{segment_data}
\alias{segment_data}
\title{Return segment data}
\usage{
segment_data(data_points)
}
\arguments{
\item{data_points}{data frame with the coordinates x and y of the points}
}
\value{
dataframe with data for a straight line between points (x, y) and x and y axis.
}
\description{
Return segment data
}
