% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{sigma_0}
\alias{sigma_0}
\title{cavitation index (Reference upstream pressure P1):}
\usage{
sigma_0(p1, p2, masl, temp)
}
\arguments{
\item{p1}{Gauge Inlet pressure (bar).}

\item{p2}{Gauge outlet pressure (bar).}

\item{masl}{meters above sea level (m).}

\item{temp}{The temperature is in Celsius.}
}
\value{
Sigma
}
\description{
The value for the operating service conditions of a valve.
}
