% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{sigma_1}
\alias{sigma_1}
\title{cavitation index (Reference downstream pressure P2):}
\usage{
sigma_1(p1, p2, masl, temp)
}
\arguments{
\item{p1}{Gauge Inlet pressure (bar)}

\item{p2}{Gauge outlet pressure (bar)}

\item{masl}{meters above sea level [m]}

\item{temp}{The temperature is in Celsius.}
}
\value{
Sigma
}
\description{
The value for the operating service conditions of a valve.
Reference downstream pressure
}
