% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{sigma_3}
\alias{sigma_3}
\title{cavitation index (Reference downstream pressure P2):}
\usage{
sigma_3(p1, p2, flow, dn, masl, temp)
}
\arguments{
\item{p1}{Gauge Inlet pressure (bar)}

\item{p2}{Gauge outlet pressure (bar)}

\item{flow}{flow in m³/s}

\item{dn}{valve diameter (m).}

\item{masl}{meters above sea level (m)}

\item{temp}{The temperature is in Celsius.}
}
\value{
sigma
}
\description{
The value for the operating service conditions of a valve.
Reference downstream pressure and Adding the cavitation caused by surface
roughness, an isolated roughness, an offset in the boundary, or by any
device for which it is not possible or convenient to evaluate a pressure
differential), the velocity head can be used in adding the DP in
the Equation.
}
