% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{Sigma_c}
\alias{Sigma_c}
\title{Constant Cavitation.}
\usage{
Sigma_c(fl)
}
\arguments{
\item{fl}{liquid pressure recovery factor}
}
\value{
Sigma_c
}
\description{
An early level of cavitation characterized by mild,
steady popping or crackling sounds that may be audible or detected by
vibration measurements. It is the next higher inflection point on the
cavitation profile above the point of incipient cavitation.
 - More regular cavitation events
 - Lower frequency sound and vibration sensed: “rumbling” sound
 - Some damage to surfaces may occur: dependent upon valve and trim styles, and materials.
}
