% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{Sigma_i}
\alias{Sigma_i}
\title{Incipient Cavitation.}
\usage{
Sigma_i(fl)
}
\arguments{
\item{fl}{liquid pressure recovery factor}
}
\value{
Sigma_i
}
\description{
The onset of cavitation, where only small vapor bubbles are
formed in the flow stream. A cavitation level sufficient to begin minor,
observable indications of pitting damage.
 - Onset of cavitation;
 - Detect using high-frequency vibration measurement;
 - Very local phenomenon; Transient: random “ticks” sound;
 - Low-level cavitation: usually not damaging;
 - Occurs prior to the loss of capacity
}
