% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{Sigma_mv}
\alias{Sigma_mv}
\title{Maximum Vibration Cavitation}
\usage{
Sigma_mv(x, b, d, e, fls)
}
\arguments{
\item{x}{valve position}

\item{b}{steepness}

\item{d}{upper value}

\item{e}{the effective dose}

\item{fls}{liquid pressure recovery full open (max between fl and Flp/Fp)}
}
\value{
Sigma_mv
}
\description{
The level of cavitation associated with peak vibration
measurements.
- Highest vibration amplitude: sounds like “marbles” or “gravel”
 - Vigorous, large scale cavitation
 - Predicted by steady flow pressure distribution (=Fl)
 - Very high damage potential
}
