% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tables.R
\name{tab_param_points}
\alias{tab_param_points}
\title{Parameters of the Points at the valve}
\usage{
tab_param_points(param_points)
}
\arguments{
\item{param_points}{tibble table}
}
\value{
param_points
}
\description{
This function prints a table of the Points parameters measured
or calculated at the valve for each type of valve.
}
