% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{type_of_flow}
\alias{type_of_flow}
\title{Type of flow}
\usage{
type_of_flow(p1, p2, temp, fl)
}
\arguments{
\item{p1}{Inlet Absolute pressure}

\item{p2}{Outlet Absolute pressure}

\item{temp}{Inlet water temperature in °C}

\item{fl}{Liquid pressure recovery factor of a control valve without attached fittings}
}
\value{
chr with "non-choked flow" or "choked flow"
}
\description{
Tested ob the flow is choked or not
}
\examples{
type_of_flow(p1 = 6.8, p2 = 2.2, temp = 89, fl = 0.9)
type_of_flow(p1 = 6.8, p2 = 2.2, temp = 89, fl = 0.6)

}
