% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpanetValve.R
\name{valve_param_calc}
\alias{valve_param_calc}
\title{data_analyze selection of the right valve type}
\usage{
valve_param_calc(base_data, valves, dn, d1, d2, masl, temp, add_factor = 1.3)
}
\arguments{
\item{base_data}{tibble table}

\item{valves}{tibble table}

\item{dn}{valve diameter (mm).}

\item{d1}{downstream pipe diameter (mm).}

\item{d2}{upstream pipe diameter (mm).}

\item{masl}{meters above sea level (m).}

\item{temp}{The water temperature is in Celsius.}

\item{add_factor}{additional factor Select valve types that meet Kvs > 1.3 Kv}
}
\value{
data_analyze tibble table
}
\description{
data_analyze selection of the right valve type
}
