% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterFunctions.R
\name{vapour_pressure}
\alias{vapour_pressure}
\title{Vapour pressure of water}
\usage{
vapour_pressure(temp = 15)
}
\arguments{
\item{temp}{is in °C}
}
\value{
Vapour pressure of water in (bar)
}
\description{
The vapor pressure of water is the pressure at which water vapor is in
thermodynamic equilibrium with its condensed state. At higher pressures water
would condense. The water vapor pressure is the partial pressure of water
vapor in any gas mixture in equilibrium with solid or liquid water.
As for other substances, water vapor pressure is a function of temperature
and can be determined with the Clausius–Clapeyron relation.
Approximation formula :
The Buck equation. where T is in °C and P is in kPa.
https://en.wikipedia.org/wiki/Vapour_pressure_of_water
}
\examples{
vapour_pressure(25)
}
\author{
Dr. Raúl Trujillo Álvarez
}
