% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{velocity}
\alias{velocity}
\title{Velocity in a pipe}
\usage{
velocity(flow, dn)
}
\arguments{
\item{flow}{cubic meter per second [m³/s]}

\item{dn}{diameter in meter [m]}
}
\value{
velocity in meter per second [m/s]
}
\description{
This function calculates the velocity of the fluid in a
  circular pipe.
}
\examples{
velocity(flow = 0.4, dn = 0.2)
}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
