% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{water_density}
\alias{water_density}
\title{Saturated water Density}
\usage{
water_density(temp)
}
\arguments{
\item{temp}{is in °C}
}
\value{
density of water in (kg/m³)
}
\description{
The density of water is about 1 gram per cubic centimetre (62 lb/cu ft):
this relationship was originally used to define the gram.
The density varies with temperature, but not linearly:
as the temperature increases, the density rises to a peak at 3.98 °C (39.16 °F)
and then decreases.
}
\examples{
water_density(14.5)

}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
