% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{zeta_vaule}
\alias{zeta_vaule}
\title{Zeta Value}
\usage{
zeta_vaule(kv, dn)
}
\arguments{
\item{kv}{Kv value}

\item{dn}{diameter in meter (m)}
}
\value{
Zeta Vaule
}
\description{
Zeta Value
}
\examples{

}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
