% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{zv_function}
\alias{zv_function}
\title{Function Resistance Coefficient Zeta}
\usage{
zv_function(x, b, d, e, zvs)
}
\arguments{
\item{x}{valve position}

\item{b}{steepness}

\item{d}{upper value}

\item{e}{the effective dose}

\item{zvs}{Resistance Coefficient full open}
}
\value{
zv Resistance Coefficient
}
\description{
Function Resistance Coefficient Zeta
}
