% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{drm_LL3}
\alias{drm_LL3}
\title{dose-response models.}
\usage{
drm_LL3(x, b, d, e)
}
\arguments{
\item{x}{valve position}

\item{b}{steepness}

\item{d}{upper value}

\item{e}{the effective dose}
}
\value{
ll3
}
\description{
Built-in dose-response models. These models are
parameterized using a unified structure with a coefficient b denoting the
steepness of the dose-response curve, d the upper asymptotes or
limits of the response, and, for some models, e the effective dose.
}
\examples{
drm_LL3(5, 2.47033, 7.85543, 3.26336)
}
