% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{fl_function}
\alias{fl_function}
\title{Function FL liquid pressure recovery factor}
\usage{
fl_function(x, b, d, e, fls)
}
\arguments{
\item{x}{valve position}

\item{b}{steepness}

\item{d}{upper value}

\item{e}{the effective dose}

\item{fls}{liquid pressure recovery full open (max between fl and Flp/Fp)}
}
\value{
fl liquid pressure recovery factor for
}
\description{
for a control valve without attached fittings
}
