% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{flow_through_valve}
\alias{flow_through_valve}
\title{Equations for in-compressible fluids}
\usage{
flow_through_valve(kv, p1, p2, temp = 15, masl = 0, fl, fr = 1)
}
\arguments{
\item{kv}{Flow coefficient in [m³/h]}

\item{p1}{Inlet Absolute pressure [bar]}

\item{p2}{Outlet Absolute pressure [bar]}

\item{temp}{temperature in [°C]}

\item{masl}{meter above sea level [m]}

\item{fl}{Liquid pressure recovery factor of a control valve without attached fittings}

\item{fr}{Reynolds number factor.}
}
\value{
Volumetric flow rate in m³/h
}
\description{
The equations listed below identify the relationships between
flow rates, flow coefficients, related installation factors, and pertinent
service conditions for control valves handling in-compressible fluids.
A sizing flow chart for in-compressible fluids is given in Annex B. of the
ISA-75.01.01-2007 : Flow Equations for Sizing Control Valves.
}
\note{
"Absolute pressure" is zero-referenced against a perfect vacuum, using
an absolute scale, so it is equal to gauge pressure plus atmospheric pressure.
"Gauge pressure" is zero-referenced against ambient air pressure, so
it is equal to absolute pressure minus atmospheric pressure. [https://en.wikipedia.org/wiki/Pressure_measurement]
}
\examples{
 flow_through_valve(kv = 238, p1 = 6.8, p2 = 2.2, temp = 80, fl = 0.6)
}
