% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{friction_zigrang}
\alias{friction_zigrang}
\title{Friction in Pipe}
\usage{
friction_zigrang(flow, dn, roughness, temp = 20)
}
\arguments{
\item{flow}{cubic meter per second (m³/s)}

\item{dn}{diameter in meter (m)}

\item{roughness}{in (mm)}

\item{temp}{temp is in °C}
}
\value{
Friction value
}
\description{
Friction calculation
}
\examples{


}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
