% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrailic_functions.R
\name{kinematic_viscosity}
\alias{kinematic_viscosity}
\title{Kinematic Viscosity}
\usage{
kinematic_viscosity(temp = 15.6)
}
\arguments{
\item{temp}{is in °C}
}
\value{
Kinematic Viscosity in (m2/s)*1e-6
}
\description{
Kinematic viscosity is the ratio of - absolute (or dynamic) viscosity
to density - a quantity in which no force is involved.
The kinematic viscosity ν is simply the dynamic viscosity μ
divided by the fluid density ρ.
The units of μ are pascals-seconds (kg/m.s) vs. the units of ν which are m2/s.
}
\examples{
kinematic_viscosity(14.5)

}
\author{
Dr. Raúl Trujillo Álvarez \email{dr.ing.trujillo@gmail.com}
}
