% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{q_max}
\alias{q_max}
\title{the maximum flow through the valve}
\usage{
q_max(p1, fl, kv, dn, d1, d2, masl, temp)
}
\arguments{
\item{p1}{Gauge upstream pressure (bar)}

\item{fl}{liquid pressure recovery factor}

\item{kv}{Flow coefficient value in (m3/h).}

\item{dn}{valve diameter (m).}

\item{d1}{downstream pipe diameter (m).}

\item{d2}{upstream pipe diameter (m).}

\item{masl}{meters above sea level (m).}

\item{temp}{The temperature is in Celsius.}
}
\value{
q_max (m3/h)
}
\description{
the maximum flow through the valve
}
