% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpanetValve.R
\name{valve_analyze}
\alias{valve_analyze}
\title{Valve Analyze}
\usage{
valve_analyze(net, report, valve_name, temperature = 15, masl = 0)
}
\arguments{
\item{net}{Network data from EPANET input file with \code{epanetReader::read.inp()}}

\item{report}{Report from EPANET report file with \code{epanetReader::read.rpt()}}

\item{valve_name}{Valve name in EPNET network and report}

\item{temperature}{Temperature is in °C}

\item{masl}{Meters above sea level [m]}
}
\value{
A Tible Table with the values of ..Values of the
}
\description{
Valve analyze using the calculation from the EPANET
}
