% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valve_hydrailic_functions.R
\name{zeta_vaule}
\alias{zeta_vaule}
\title{Resistance Coefficient Zeta in function of Kv}
\usage{
zeta_vaule(dn, kv)
}
\arguments{
\item{dn}{diameter in millimeter (mm)}

\item{kv}{Kv Flow coefficient in m³/h}
}
\value{
Zeta Vaule
}
\description{
Pressure drop or head loss is proportional to the velocity
in valves or fittings. For the most engineering practices it can be assumed
that pressure drop or head loss due to flow of fluids in turbulent range
through valves and fittings is proportional to square of velocity.
}
\examples{
zeta_vaule( dn = 500, kv = 7247.229)

}
\author{
Dr. Raúl Trujillo Álvarez
}
